Feature: add transaction many (#60)
* fix: use parseInt for amount and balance to ensure integer values

Changed amount and balance parsing from parseFloat to parseInt to prevent decimal values and enforce integer-only inputs for transactions and account setup. This improves data consistency and validation.

* feat(transaction): add quantity selector and batch transaction creation

Added quantity controls to allow users to create multiple transactions at once. Updated logic to batch transaction creation using createTransactionMany for improved efficiency and user experience.

* refactor(utils): add arrayFill utility and refactor transaction creation

Introduce arrayFill function to generate arrays of specified length and value. Refactor transaction creation to use arrayFill for improved clarity. Add unit tests for arrayFill to ensure correct behavior.

* refactor(env): make transaction batch size configurable via env variable

Added EXPO_PUBLIC_TRANSACTION_INSERT_MAX_BATCH_SIZE to .env.example and env.ts to allow dynamic configuration of max batch size for transaction inserts. Updated add-transaction.tsx to use this value for improved flexibility.

+++ src/features/account-balance-setup-modal/index.tsx
@@ -62,7 +62,7 @@ export function AccountBalanceSetupModal({
       userId: userId,
       name: "กระเป๋าสตางค์",
       type: "cash",
-      balance: parseFloat(balance),
+      balance: parseInt(balance),
       currency: "THB",
       icon: "💵",
     });
